Refactor QuickCache runtime for maintainability
